Add simple wildcarding to server ACL targets.
There have been several requests and patches proposed to add support for
some kind of extended wildcard matching in ACLs for kadmind.

This patch adds code that checks for an asterisk at the beginning of
the target, but followed by a non-digit.  It then matches on the rest
of the target.

Hence this ACL ...

*/dept.admin@TEST.REALM * */*.dept.test.realm@TEST.REALM

... allows all admin principals from Dept "dept" to manage instance
principals for hosts in their DNS domain in our central realm.  This has
freed us up from manually creating thousands of principals on behalf of
departments, or writing a complex devolved web front end.